back/ML/requirements.txt
fastapi
uvicorn
python-multipart
google-cloud-vision
pdf2image
pillow
pytesseract
